requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
plotly>=5.18.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
arch>=6.0.0
TA-Lib>=0.4.28
